pyproject.toml
[tool.poetry]
name = "civicsentinel"
version = "0.1.0"
description = "AI-powered CCTV surveillance system backend"
authors = ["CivicSentinel Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.35"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-multipart = "^0.0.12"
pillow = "^11.0.0"
ultralytics = "^8.3.0"
opencv-python = "^4.10.0"
shapely = "^2.0.6"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
black = "^24.10.0"
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
